refactor the code to support future changes
