+++ src/data/update/compress_reference_dict.js
@@ -1,0 +1,29 @@
+export const COURSE_CODE_ENCODE_DICT = {
+  "SWE": 0,
+  "STS": 1,
+  "PHY": 2,
+  "MGT": 3,
+  "MAT": 4,
+  "LIB": 5,
+  "FRL": 6,
+  "ENG": 7,
+  "ECE": 8,
+  "CSE": 9,
+  "CHY": 10,
+};
+
+export const COURSE_SLOT_ENCODE_DICT = {
+  "L9+L10": 0,
+  "L35+L36": 1,
+  "L31+L32": 2,
+  "L21+L22": 3,
+  "L19+L20": 4,
+  "L1+L2": 5,
+  "E+TE": 6,
+  "D+TD": 7,
+  "D": 8,
+  "C+TC": 9,
+  "C": 10,
+  "B+TB": 11,
+  "A+TA": 12,
+};
