+++ src/data/update/time_table.js
@@ -1,0 +1,69 @@
+export const time_table = {
+  "Monday": [
+    "L1",
+    "L2",
+    "TA+L3",
+    "TB+L4",
+    "TC+L25",
+    "TD+L26",
+    "TE+L27",
+    "L28"
+  ],
+  "Tuesday": [
+    "C+L5",
+    "D+L6",
+    "E+L7",
+    "L8",
+    "A+L29",
+    "B+L30",
+    "L31",
+    "L32"
+  ],
+  "Wednesday": [
+    "L9",
+    "TE+L10",
+    "TB+L11",
+    "TA+L12",
+    "TC+L33",
+    "TD+L34",
+    "L35",
+    "L36"
+  ],
+  "Thursday": [
+    "E+L13",
+    "L14",
+    "D+L15",
+    "C+L16",
+    "B+L37",
+    "A+L38",
+    "L39",
+    "L40"
+  ],
+  "Friday": [
+    "A+L17",
+    "B+L18",
+    "L19",
+    "L20",
+    "D+L41",
+    "C+L42",
+    "E+L43",
+    "L44"
+  ],
+  "Saturday": [
+    "L21",
+    "L22",
+    "C+L23",
+    "D+L24",
+    "A+L45",
+    "B+L46",
+    "E+L47",
+    "L48"
+  ]
+};
+
+
+/**
+ * @type {string[]}
+ */
+export const time_arr = ["9:00", "10:00", "11:00", "12:00", "2:00", "3:00", "4:00", "5:00"];
+
